requirements.txt
python-telegram-bot
pandas
pydantic
gspread
oauth2client
pydantic-settings
email-validator